Assignment 2
Played around with the code and Googled, but could not figure out how to add the Facebook, Twitter and Instagram icons or "Email Address" and Keep In Touch buttons. The footer was the most challenging for me. I also tried to use div, classes and/or ID to consolidate my code, but the site seemed to come together without them. Would love to go over!