Uploading Homework 3 (Monument Lab Starter)
I'm observing that using block level semantic elements such as h2 h3 h4 to create padding creates problems later, like after you've added flexbox and are creating white space. 

I'd like to learn more about ::before and ::after and how these are used in transitions. 

I used a grid to implement the gallery portion of the project — but I was having the devil's time getting the gutter spaces between the images to be even. Wondering ways to address this...?