Add files via upload
Homework submitted late on 08/28/17

This website will be about my backpacking adventures through our national parks.  I'd like to showcase my photos with descriptions of the pictures like where it took place and how I got there.  

My header/nav will be centered at the top of the page and not be stickied because I don't want it to follow as you scroll down.  There will be 3 links "home, "about," and "contact."  Under this will be a large area with an image; however I'd like the image to change and fade into a different image every 2-3 seconds.

The main content of the page will be 6 pictures with 6 paragraphs to describe the pictures.  I'm thinking about having something to click to expand the text rather than having the text always showing.  I think this will depend on how it all looks aesthetically.  Also I don't want navigation of the site to be confusing so I will have to take that into consideration.  

When looked at on a larger screen, the pictures and text will be in col