Add files via upload
This most likely isn't how my final project will look, it's more like a rough draft to give me some ideas and guidance.  The layout I drew will allow me to get started creating the structure of my website.  One of my hobbies is backpacking and I have quite a few pictures from my backpacking trips so I think my website will be about that.  It will include pictures of where I've been and the content will be about the various national parks as well as tips for people planning a trip.

Something I'd like to do, which wasn't yet covered in the course, is have a picture slide show towards the top where the background image usually is.  The main content of my site will be 2 rows.  Row 1 will have 3 pictures with 3 text, and row 2 will have the same.  I'd like to have some hover effects such as the image slightly zooming in or image opacity.

Since we studied forms, I'd like to add a contact form which would allow the user to send me a message if they have any questions. This will be in the